infrastructure/requirements.txt
azure-cosmos>=4.5.0
azure-identity>=1.15.0